Dynamic buttons
The buttons for selecting the question filters for creating a random test have been made dynamic so that they query the database and pull all distinct values in the applicable tables to create buttons. Created a route in devapp.py to call the create_random_test.html and query the database to get the data create the buttons.