Update to test creation.
While this is the most current update to this branch no pulls should NOT be made to the master as it currently is, as substantial modifications have been made and the js for manual test creation has not been updated accordingly. This is unstable and requires major testing and debugging. Random testing with perfect conditions is the only thing I know is functional at this time.